+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [unreleased]
 
+ * Core: The health and readiness probes return now a JSON with the actual errors
+
 ## 2.1.1 (June 13, 2022)
 
  * Admin Toolbar: Fixed validation of route '/'
@@ -11,7 +13,7 @@
 
  * Portal: Re-check authentication expiration at least every 60sec, so, if the session for some reason expires (or gets revoked)
    the user will get notified faster.
- * Dropped Node.js 12.x support
+ * Core: Dropped Node.js 12.x support
  * Portal: Prevent a loop if error messages can not be sent to the server
  * MongoDB Session Provider: **BREAKING CHANGE**: Changed config structure to be able to pass parameters to connect-mongo,
    such as *ttl* and *autoRemove*.